migrate helper funcions
